Separate function to parse BuildCfg into Options
License: MIT
